pyproject.toml
[project]
name = "auto-bond-purchase"
version = "0.1.0"
description = ""
authors = [
    {name = "jun14111", email = "1411187249@qq.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.50.0",
    "requests",
    "python-dotenv",
    "tensorflow>=2.15.0",
    "opencv-python>=4.9.0",
    "numpy>=1.23.5,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
